Bulge mass calculation: changes in model_mergers.c to better track secular and merger triggered bulge growth. Additional radio mode check: additional check in model_cooling.c just in case Mvir=0. Fix for dT: accidentally stored as int instead of float in allvars.h and the python scripts. Parameter files: minor changes in to the comments.

+++ code/core_allvars.h
@@ -49,7 +49,7 @@ struct GALAXY_OUTPUT
   int   mergeType;  //0=none; 1=minor merger; 2=major merger; 3=disk instability; 4=disrupt to ICS
   int   mergeIntoID;
   int   mergeIntoSnapNum;
-  int   dT;
+  float   dT;
 
   // properties of subhalo at the last time this galaxy was a central galaaxy 
   float Pos[3];
@@ -115,7 +115,7 @@ struct GALAXY
   int   mergeType;  //0=none; 1=minor merger; 2=major merger; 3=disk instability; 4=disrupt to ICS
   int   mergeIntoID;
   int   mergeIntoSnapNum;
-  int   dT;
+  float   dT;
 
   // properties of subhalo at the last time this galaxy was a central galaxy 
   float Pos[3];
